requirements.txt
basicsr ==1.4.2
nibabel ==4.0.2
ninja ==1.11.1
opencv-python ==4.7.0.68
pandas ==1.3.5
pydicom ==2.3.1
pytorch-msssim ==0.2.1
pyyaml ==6.0
runstats ==2.0.0
scikit-image ==0.19.3
scikit-learn ==1.0.2
scipy ==1.7.3
simpleitk ==2.3.0
statsmodels ==0.13.5
tensorboard ==2.11.2
timm ==0.6.12
torch ==1.13.1
torchio ==0.18.92
torchmetrics ==0.7.0
torchvision ==0.14.1
tqdm ==4.64.1
lpips==0.1.4
matplotlib == 3.7.5
einops == 0.8.0
mpmath == 1.3.0

